backend/requirements.txt
# FastAPI & Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Supabase
supabase==2.0.3
postgrest-py==0.13.2

# Database
psycopg2-binary==2.9.9

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP & CORS
httpx==0.25.2

# Utils
python-dateutil==2.8.2

